Refactor comand structure in auth
