--- crates/bitwarden/src/auth/commands/mod.rs
@@ -1,3 +1,0 @@
-mod login;
-
-pub(crate) use login::*;
