+++ crates/bitwarden/src/auth/mod.rs
@@ -1,4 +1,5 @@
 pub(super) mod api;
-pub mod commands;
+pub mod login;
 pub mod request;
 pub mod response;
+pub mod renew;
