+++ crates/bitwarden/src/client/client.rs
@@ -4,10 +4,7 @@ use reqwest::header::{self};
 use uuid::Uuid;
 
 use crate::{
-    auth::{
-        commands::{access_token_login, renew_token},
-        response::ApiKeyLoginResponse,
-    },
+    auth::{login::access_token_login, renew::renew_token, response::ApiKeyLoginResponse},
     client::{
         client_settings::{ClientSettings, DeviceType},
         encryption_settings::{EncryptionSettings, SymmetricCryptoKey},
@@ -22,7 +19,7 @@ use crate::auth::request::AccessTokenLoginRequest;
 use {
     crate::{
         auth::{
-            commands::{api_key_login, password_login},
+            login::{api_key_login, password_login},
             request::{ApiKeyLoginRequest, PasswordLoginRequest},
             response::PasswordLoginResponse,
         },
